Renaming ComposableTechnique -> ComposableEvolutionaryTechnique.
Also pulling the class out of techniques.py
