+++ docs/source/techniques.rst
@@ -1,4 +1,4 @@
-.. currentmodule:: opentuner.search.technique
+.. currentmodule:: opentuner.search.composableevolutionarytechniques
 
 ********************
 Current Techniques
@@ -10,27 +10,27 @@ OpenTuner has a library of existing search techniques.
 Composable Search Techniques
 =================================
 
-A ComposableSearchTechnique allows for composition between the search technique and any operators. Creating a ComposableSearchTechnique requires implementing 3 methods:
+A ComposableEvolutionaryTechnique allows for composition between the search technique and any operators. Creating a ComposableEvolutionaryTechnique requires implementing 3 methods:
 
- * :meth:`minimum_number_of_parents <ComposableSearchTechnique.minimum_number_of_parents>`
- * :meth:`get_parents <ComposableSearchTechnique.get_parents>`
- * :meth:`update_population <ComposableSearchTechnique.update_population>`
+ * :meth:`minimum_number_of_parents <ComposableEvolutionaryTechnique.minimum_number_of_parents>`
+ * :meth:`get_parents <ComposableEvolutionaryTechnique.get_parents>`
+ * :meth:`update_population <ComposableEvolutionaryTechnique.update_population>`
 
 Additionally, the following methods may be overridden for further customization
 
- * :meth:`make_population_member <ComposableSearchTechnique.make_population_member>`
- * :meth:`select_parameters <ComposableSearchTechnique.select_parameters>`
- * :meth:`get_default_operator <ComposableSearchTechnique.get_default_operator>`
+ * :meth:`make_population_member <ComposableEvolutionaryTechnique.make_population_member>`
+ * :meth:`select_parameters <ComposableEvolutionaryTechnique.select_parameters>`
+ * :meth:`get_default_operator <ComposableEvolutionaryTechnique.get_default_operator>`
 
 The following methods are useful when choosing parents or updating the population:
 
- * :meth:`lt <ComposableSearchTechnique.lt>`
- * :meth:`lte <ComposableSearchTechnique.lte>`
- * :meth:`get_global_best_configuration <ComposableSearchTechnique.get_global_best_configuration>`
+ * :meth:`lt <ComposableEvolutionaryTechnique.lt>`
+ * :meth:`lte <ComposableEvolutionaryTechnique.lte>`
+ * :meth:`get_global_best_configuration <ComposableEvolutionaryTechnique.get_global_best_configuration>`
 
-A ComposableSearchTechnique will yields configurations generated by successive iterations of applying operators on the configurations returned by :meth:`get_parents <ComposableSearchTechnique.get_parents>` and updating the population with the new configuration through :meth:`update_population <ComposableSearchTechnique.update_population>`
+A ComposableEvolutionaryTechnique will yields configurations generated by successive iterations of applying operators on the configurations returned by :meth:`get_parents <ComposableEvolutionaryTechnique.get_parents>` and updating the population with the new configuration through :meth:`update_population <ComposableEvolutionaryTechnique.update_population>`
 
-.. autoclass:: ComposableSearchTechnique
+.. autoclass:: ComposableEvolutionaryTechnique
 
 	.. automethod:: minimum_number_of_parents
